create classes for shunts
